\name{DateTime-class}
\Rdversion{1.1}
\docType{class}
\alias{DateTime-class}

\title{Class "DateTime"}
\description{A class union of \code{"Date"} and \code{"POSIXct"} classes.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  No methods defined with class "DateTime" in the signature.
}
\seealso{
  \code{\link{WqData-class}}
}
\examples{
showClass("DateTime")
}
\keyword{classes}
