% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapYear.R
\name{leapYear}
\alias{leapYear}
\title{leapYear}
\usage{
leapYear(x)
}
\arguments{
\item{x}{integer year}
}
\description{
\code{TRUE} if \code{x} is a leap year, \code{FALSE}
otherwise.
}
\author{
Alan Jassby, James Cloern
}
