% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSub.R
\name{meanSub}
\alias{meanSub}
\title{meanSub}
\usage{
meanSub(x, sub, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{sub}{integer index}

\item{na.rm}{logical}
}
\description{
meanSub
}
\author{
Alan Jassby, James Cloern
}
