% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wql-package.R
\docType{methods}
\name{phenoAmp-methods}
\alias{phenoAmp-methods}
\title{Methods for Function phenoAmp}
\description{
Finds various measures of the amplitude of the annual cycle.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ts\")")}{See
\code{\link{phenoAmp,ts-method}}} \item{list("signature(x = \"zoo\")")}{See
\code{\link{phenoAmp,zoo-method}}} }
}

\keyword{methods}
