\name{phenoPhase-methods}
\docType{methods}
\alias{phenoPhase-methods}

\title{Methods for Function phenoPhase}

\description{
Finds various measures of the phase of the annual cycle.
}

\section{Methods}{
  \describe{
    \item{\code{signature(x = "ts")}}{See \code{\link{phenoPhase,ts-method}}}
    \item{\code{signature(x = "zoo")}}{See \code{\link{phenoPhase,zoo-method}}}
  }
}

\keyword{methods}
