% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2pss.R
\name{R2pss}
\alias{R2pss}
\title{R2pss}
\usage{
R2pss(R, t, p = 0)
}
\arguments{
\item{R}{conductivity ratio, dimensionless}

\item{t}{temperature, Celsius}

\item{p}{gauge pressure, decibar}
}
\description{
R2pss
}
\author{
Alan Jassby, James Cloern
}
