% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wql-package.R
\docType{methods}
\name{tsMake-methods}
\alias{tsMake-methods}
\title{Methods for Function tsMake}
\description{
Creates a matrix of observations indexed by time.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"WqData\")")}{See
\code{\link{tsMake,WqData-method}}} }
}

\keyword{methods}
