% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wql-package.R
\docType{class}
\name{zoo-class}
\alias{zoo-class}
\title{Class "zoo"}
\description{
Registration of S3 class \code{"zoo"} as a formally defined class. Used here
to allow the \code{"zoo"} class to appear in method signatures.
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}

\examples{

showClass("zoo")

}
\seealso{
\code{\link{phenoAmp}}, \code{\link{phenoPhase}}
}
\keyword{classes}
