% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wql-package.R
\docType{class}
\name{DateTime-class}
\alias{DateTime-class}
\title{Class "DateTime"}
\description{
A class union of \code{"Date"} and \code{"POSIXct"} classes.
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}

\examples{

showClass("DateTime")

}
\seealso{
\code{\link{WqData-class}}
}
\keyword{classes}
