\name{phenoAmp-methods}
\docType{methods}
\alias{phenoAmp-methods}

\title{Methods for Function phenoAmp}

\description{
Finds various measures of the amplitude of the annual cycle.
}

\section{Methods}{
  \describe{
    \item{\code{signature(x = "ts")}}{See \code{\link{phenoAmp,ts-method}}}
    \item{\code{signature(x = "zoo")}}{See \code{\link{phenoAmp,zoo-method}}}
  }
}

\keyword{methods}