% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wql-package.R
\docType{methods}
\name{phenoPhase-methods}
\alias{phenoPhase-methods}
\title{Methods for Function phenoPhase}
\description{
Finds various measures of the phase of the annual cycle.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ts\")")}{See
\code{\link{phenoPhase,ts-method}}} \item{list("signature(x =
\"zoo\")")}{See \code{\link{phenoPhase,zoo-method}}} }
}

\keyword{methods}
