\name{tsMake-methods}
\docType{methods}
\alias{tsMake-methods}

\title{Methods for Function tsMake}

\description{
Creates a matrix of observations indexed by time.
}

\section{Methods}{
  \describe{
    \item{\code{signature(x = "WqData")}}{See \code{\link{tsMake,WqData-method}}}
  }
}

\keyword{methods}