\name{wq-package}
\alias{wq-package}
\alias{wq}
\docType{package}
\title{
Exploring water quality monitoring data
}
\description{
Functions to assist in the processing and exploration of data from monitoring programs for aquatic ecosystems. The name "wq" stands for "water quality" and reflects a focus on time series data for physical and chemical properties of water, as well as the plankton. The package is intended for programs that sample approximately monthly at discrete stations, a feature of many legacy data sets. Although our emphasis is on aquatic ecosystems, most of the functions should be useful for analysis of similar-frequency time series regardless of the subject matter.
}
\details{
\tabular{ll}{
Package: \tab wq\cr
Type: \tab Package\cr
Version: \tab 0.3-5\cr
Date: \tab 2012-03-01\cr
Depends: \tab R (>= 2.14), methods, zoo\cr
Suggests: \tab ggplot2 (>= 0.9), grid, reshape2\cr
License: \tab GPL-2\cr
LazyData: \tab yes\cr
}
}
\author{
Alan D. Jassby and James E. Cloern

Maintainer: Alan Jassby <adjassby@ucdavis.edu>
}
\keyword{ package }
