\name{wq-package}
\alias{wq-package}
\alias{wq}
\docType{package}
\title{
Exploring water quality monitoring data
}
\description{
Functions to assist in the processing and exploration of data from monitoring programs for aquatic ecosystems. The focus is on time series data for physical and chemical properties of water, as well as the plankton. The package is intended for programs that sample approximately monthly at discrete stations.
}
\details{
\tabular{ll}{
Package: \tab wq\cr
Type: \tab Package\cr
Version: \tab 0.2-9\cr
Date: \tab 2010-12-30\cr
Depends: \tab R (>= 2.11.0), methods, zoo\cr
Suggests: \tab ggplot2, reshape\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
Alan D. Jassby and James E. Cloern

Maintainer: Alan Jassby <adjassby@ucdavis.edu>
}
\keyword{ package }
