\name{zoo-class}
\Rdversion{1.1}
\docType{class}
\alias{zoo-class}

\title{Class "zoo"}
\description{Registration of S3 class \code{"zoo"} as a formally defined class. Used here to allow the \code{"zoo"} class to appear in method signatures.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"}}
  }
}

\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}

\seealso{
\code{\link{phenoAmp}}, \code{\link{phenoPhase}}
}

\examples{
showClass("zoo")
}

\keyword{classes}
